--- 방예혁/src/main/java/com/gdsc/jwtpractice/JwtPracticeApplication.java
@@ -1,13 +1,0 @@
-package com.gdsc.jwtpractice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JwtPracticeApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(JwtPracticeApplication.class, args);
-    }
-
-}
