--- 방예혁/src/main/java/com/gdsc/jwtpractice/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.gdsc.jwtpractice.repository;
-
-import com.gdsc.jwtpractice.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByMemberId(String username);
-}
