--- 방예혁/src/main/java/com/gdsc/jwtpractice/service/JwtUserDetailsService.java
@@ -1,36 +1,0 @@
-package com.gdsc.jwtpractice.service;
-
-import com.gdsc.jwtpractice.domain.Member;
-import com.gdsc.jwtpractice.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-// UserDetailsService 인터페이스를 구현한 클래스
-@Service
-@RequiredArgsConstructor
-public class JwtUserDetailsService implements UserDetailsService {
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    // 전달받은 username 값을 통해 memberRepository에서 유저를 찾아서 UserDetails 객체로 리턴
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return memberRepository.findByMemberId(username)
-                .map(this::createUserDetails)
-                .orElseThrow(() -> new UsernameNotFoundException("해당하는 유저를 찾을 수 없습니다."));
-    }
-
-    // DB에 해당하는 User 값이 존재한다면 UserDetails 객체로 리턴
-    private UserDetails createUserDetails(Member member) {
-        return User.builder()
-                .username(member.getUsername())
-                .password(passwordEncoder.encode(member.getPassword()))
-                .roles(member.getRoles().toArray(new String[0]))
-                .build();
-    }
-}
