--- 방예혁/src/test/java/com/gdsc/jwtpractice/JwtPracticeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.gdsc.jwtpractice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class JwtPracticeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
