+++ 이현복/src/main/java/gdsc/skhu/jwt/JwtApplication.java
@@ -1,0 +1,13 @@
+package gdsc.skhu.jwt;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class JwtApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(JwtApplication.class, args);
+    }
+
+}
