+++ 이현복/src/main/java/gdsc/skhu/jwt/domain/DTO/LoginDTO.java
@@ -1,0 +1,9 @@
+package gdsc.skhu.jwt.domain.DTO;
+
+import lombok.Data;
+
+@Data
+public class LoginDTO {
+    private String memberId;
+    private String password;
+}
