+++ 이현복/src/main/java/gdsc/skhu/jwt/service/MemberService.java
@@ -1,0 +1,37 @@
+package gdsc.skhu.jwt.service;
+
+import gdsc.skhu.jwt.domain.DTO.TokenDTO;
+import gdsc.skhu.jwt.jwt.TokenProvider;
+import gdsc.skhu.jwt.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+    private final TokenProvider tokenProvider;
+    private final AuthenticationManagerBuilder authenticationManagerBuilder;
+
+    @Transactional
+    public TokenDTO login(String memberId, String password) {
+        // 1. ID/PW 를 기반으로 Authentication 객체 생성
+        // 이때 authentication 객체는 인증 여부를 확인하는 authenticated 값이 false
+        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(memberId, password);
+
+        // 2. 실제 검증(사용자 비밀번호 체크)이 이루어지는 부분
+        // authenticate 매서드가 실행될 때 JwtUserDetailsService 에서 만든 loadUserByUsername 메서드가 실행
+        Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
+
+        // 3. 인증된 정보를 기반으로 JWT 토큰 생성
+        TokenDTO tokenDTO = tokenProvider.createToken(authentication);
+
+        return tokenDTO;
+    }
+}
