+++ 이현복/src/main/resources/data.sql
@@ -1,0 +1,4 @@
+use jwt;
+
+insert into member(member_id, password) values ('admin', '1234');
+insert into member_roles(member_member_id, roles) values ('admin', 'ADMIN');
