+++ 이현복/src/test/java/gdsc/skhu/jwt/JwtApplicationTests.java
@@ -1,0 +1,13 @@
+package gdsc.skhu.jwt;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class JwtApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
